Updated UI to display entire system functionality
The previous UI was only displaying how to register a property. GUI has been fully redesigned to now have a home screen on the left side of the screen. The selected category of either "Register a Property, Manage Properties, View Payments or Department of Environment" will then show up on the right side. Users will have the ability to switch between different modes easily with the UI overhaul.